Changing image with a button
The button allows you to buy now rooms and it decreses the cost from your money. Dosen't need any additional written code when adding a new room image in the textures array.
