+++ Assets/Textures/Room 2.png.meta
